UI: Fix react being bundled in service-worker.js build
Because the service-worker.js file was indirectly importing react.
